backend/requirements.txt
# Django
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0

# Database & Task Queue
django-q2>=1.6

# Image Processing
Pillow>=10.0

# ASGI Server
uvicorn[standard]>=0.27
gunicorn>=21.0

# LLM Integration
langchain>=0.3
langchain-community>=0.3
langchain-ollama>=0.2
langchain-openai>=0.2
openai>=1.0
